Extract resolve area to a separate util
